change BM file structure - edits
